+++ cli/tests/stake.rs
@@ -11,9 +11,7 @@ use {
     solana_cli_output::{parse_sign_only_reply_string, OutputFormat},
     solana_faucet::faucet::run_local_faucet,
     solana_rpc_client::rpc_client::RpcClient,
-    solana_rpc_client_api::{
-        request::DELINQUENT_VALIDATOR_SLOT_DISTANCE, response::StakeActivationState,
-    },
+    solana_rpc_client_api::request::DELINQUENT_VALIDATOR_SLOT_DISTANCE,
     solana_rpc_client_nonce_utils::blockhash_query::{self, BlockhashQuery},
     solana_sdk::{
         account_utils::StateMut,
@@ -28,275 +26,12 @@ use {
         stake::{
             self,
             instruction::LockupArgs,
-            state::{Delegation, Lockup, StakeActivationStatus, StakeAuthorize, StakeStateV2},
+            state::{Lockup, StakeAuthorize, StakeStateV2},
         },
-        sysvar::stake_history::{self, StakeHistory},
     },
     solana_streamer::socket::SocketAddrSpace,
     solana_test_validator::{TestValidator, TestValidatorGenesis},
 };
-
-#[test]
-fn test_stake_redelegation() {
-    let mint_keypair = Keypair::new();
-    let mint_pubkey = mint_keypair.pubkey();
-    let authorized_withdrawer = Keypair::new().pubkey();
-    let faucet_addr = run_local_faucet(mint_keypair, None);
-
-    let slots_per_epoch = 32;
-    let test_validator = TestValidatorGenesis::default()
-        .fee_rate_governor(FeeRateGovernor::new(0, 0))
-        .rent(Rent {
-            lamports_per_byte_year: 1,
-            exemption_threshold: 1.0,
-            ..Rent::default()
-        })
-        .epoch_schedule(EpochSchedule::custom(
-            slots_per_epoch,
-            slots_per_epoch,
-            /* enable_warmup_epochs = */ false,
-        ))
-        .faucet_addr(Some(faucet_addr))
-        .start_with_mint_address(mint_pubkey, SocketAddrSpace::Unspecified)
-        .expect("validator start failed");
-
-    let rpc_client =
-        RpcClient::new_with_commitment(test_validator.rpc_url(), CommitmentConfig::processed());
-    let default_signer = Keypair::new();
-
-    let mut config = CliConfig::recent_for_tests();
-    config.json_rpc_url = test_validator.rpc_url();
-    config.signers = vec![&default_signer];
-
-    request_and_confirm_airdrop(
-        &rpc_client,
-        &config,
-        &config.signers[0].pubkey(),
-        100_000_000_000,
-    )
-    .unwrap();
-
-    // Create vote account
-    let vote_keypair = Keypair::new();
-    config.signers = vec![&default_signer, &vote_keypair];
-    config.command = CliCommand::CreateVoteAccount {
-        vote_account: 1,
-        seed: None,
-        identity_account: 0,
-        authorized_voter: None,
-        authorized_withdrawer,
-        commission: 0,
-        sign_only: false,
-        dump_transaction_message: false,
-        blockhash_query: BlockhashQuery::All(blockhash_query::Source::Cluster),
-        nonce_account: None,
-        nonce_authority: 0,
-        memo: None,
-        fee_payer: 0,
-        compute_unit_price: None,
-    };
-    process_command(&config).unwrap();
-
-    // Create second vote account
-    let vote2_keypair = Keypair::new();
-    config.signers = vec![&default_signer, &vote2_keypair];
-    config.command = CliCommand::CreateVoteAccount {
-        vote_account: 1,
-        seed: None,
-        identity_account: 0,
-        authorized_voter: None,
-        authorized_withdrawer,
-        commission: 0,
-        sign_only: false,
-        dump_transaction_message: false,
-        blockhash_query: BlockhashQuery::All(blockhash_query::Source::Cluster),
-        nonce_account: None,
-        nonce_authority: 0,
-        memo: None,
-        fee_payer: 0,
-        compute_unit_price: None,
-    };
-    process_command(&config).unwrap();
-
-    // Create stake account
-    let stake_keypair = Keypair::new();
-    config.signers = vec![&default_signer, &stake_keypair];
-    config.command = CliCommand::CreateStakeAccount {
-        stake_account: 1,
-        seed: None,
-        staker: None,
-        withdrawer: None,
-        withdrawer_signer: None,
-        lockup: Lockup::default(),
-        amount: SpendAmount::Some(50_000_000_000),
-        sign_only: false,
-        dump_transaction_message: false,
-        blockhash_query: BlockhashQuery::All(blockhash_query::Source::Cluster),
-        nonce_account: None,
-        nonce_authority: 0,
-        memo: None,
-        fee_payer: 0,
-        from: 0,
-        compute_unit_price: None,
-    };
-    process_command(&config).unwrap();
-
-    // Delegate stake to `vote_keypair`
-    config.signers = vec![&default_signer];
-    config.command = CliCommand::DelegateStake {
-        stake_account_pubkey: stake_keypair.pubkey(),
-        vote_account_pubkey: vote_keypair.pubkey(),
-        stake_authority: 0,
-        force: true,
-        sign_only: false,
-        dump_transaction_message: false,
-        blockhash_query: BlockhashQuery::default(),
-        nonce_account: None,
-        nonce_authority: 0,
-        memo: None,
-        fee_payer: 0,
-        redelegation_stake_account: None,
-        compute_unit_price: None,
-    };
-    process_command(&config).unwrap();
-
-    // wait for new epoch plus one additional slot for rewards payout
-    wait_for_next_epoch_plus_n_slots(&rpc_client, 1);
-
-    let check_activation_status = |delegation: &Delegation,
-                                   expected_state: StakeActivationState,
-                                   expected_active_stake: u64| {
-        let stake_history_account = rpc_client.get_account(&stake_history::id()).unwrap();
-        let stake_history: StakeHistory =
-            solana_sdk::account::from_account(&stake_history_account).unwrap();
-        let current_epoch = rpc_client.get_epoch_info().unwrap().epoch;
-        let StakeActivationStatus {
-            effective,
-            activating,
-            deactivating,
-        } = delegation.stake_activating_and_deactivating(current_epoch, &stake_history, None);
-        let stake_activation_state = if deactivating > 0 {
-            StakeActivationState::Deactivating
-        } else if activating > 0 {
-            StakeActivationState::Activating
-        } else if effective > 0 {
-            StakeActivationState::Active
-        } else {
-            StakeActivationState::Inactive
-        };
-        assert_eq!(stake_activation_state, expected_state);
-        assert_eq!(effective, expected_active_stake);
-    };
-
-    // `stake_keypair` should now be delegated to `vote_keypair` and fully activated
-    let stake_account = rpc_client.get_account(&stake_keypair.pubkey()).unwrap();
-    let stake_state: StakeStateV2 = stake_account.state().unwrap();
-
-    let rent_exempt_reserve = match stake_state {
-        StakeStateV2::Stake(meta, stake, _) => {
-            assert_eq!(stake.delegation.voter_pubkey, vote_keypair.pubkey());
-            check_activation_status(
-                &stake.delegation,
-                StakeActivationState::Active,
-                50_000_000_000 - meta.rent_exempt_reserve,
-            );
-            meta.rent_exempt_reserve
-        }
-        _ => panic!("Unexpected stake state!"),
-    };
-
-    check_balance!(50_000_000_000, &rpc_client, &stake_keypair.pubkey());
-
-    let stake2_keypair = Keypair::new();
-
-    // Add an extra `rent_exempt_reserve` amount into `stake2_keypair` before redelegation to
-    // account for the `rent_exempt_reserve` balance that'll be pealed off the stake during the
-    // redelegation process
-    request_and_confirm_airdrop(
-        &rpc_client,
-        &config,
-        &stake2_keypair.pubkey(),
-        rent_exempt_reserve,
-    )
-    .unwrap();
-
-    // wait for a new epoch to ensure the `Redelegate` happens as soon as possible (i.e. till the
-    // last reward distribution block in the new epoch) to reduce the risk of a race condition
-    // when checking the stake account correctly enters the deactivating state for the
-    // remainder of the current epoch.
-    wait_for_next_epoch_plus_n_slots(&rpc_client, 1);
-
-    // Redelegate to `vote2_keypair` via `stake2_keypair
-    config.signers = vec![&default_signer, &stake2_keypair];
-    config.command = CliCommand::DelegateStake {
-        stake_account_pubkey: stake_keypair.pubkey(),
-        vote_account_pubkey: vote2_keypair.pubkey(),
-        stake_authority: 0,
-        force: true,
-        sign_only: false,
-        dump_transaction_message: false,
-        blockhash_query: BlockhashQuery::default(),
-        nonce_account: None,
-        nonce_authority: 0,
-        memo: None,
-        fee_payer: 0,
-        redelegation_stake_account: Some(1),
-        compute_unit_price: None,
-    };
-    process_command(&config).unwrap();
-
-    // `stake_keypair` should now be deactivating
-    let stake_account = rpc_client.get_account(&stake_keypair.pubkey()).unwrap();
-    let stake_state: StakeStateV2 = stake_account.state().unwrap();
-    let StakeStateV2::Stake(_, stake, _) = stake_state else {
-        panic!()
-    };
-    check_activation_status(
-        &stake.delegation,
-        StakeActivationState::Deactivating,
-        50_000_000_000 - rent_exempt_reserve,
-    );
-
-    // `stake_keypair2` should now be activating
-    let stake_account = rpc_client.get_account(&stake2_keypair.pubkey()).unwrap();
-    let stake_state: StakeStateV2 = stake_account.state().unwrap();
-    let StakeStateV2::Stake(_, stake, _) = stake_state else {
-        panic!()
-    };
-    check_activation_status(&stake.delegation, StakeActivationState::Activating, 0);
-
-    // check that all the stake, save `rent_exempt_reserve`, have been moved from `stake_keypair`
-    // to `stake2_keypair`
-    check_balance!(rent_exempt_reserve, &rpc_client, &stake_keypair.pubkey());
-    check_balance!(50_000_000_000, &rpc_client, &stake2_keypair.pubkey());
-
-    // wait for new epoch plus reward blocks
-    wait_for_next_epoch_plus_n_slots(&rpc_client, 1);
-
-    // `stake_keypair` should now be deactivated
-    let stake_account = rpc_client.get_account(&stake_keypair.pubkey()).unwrap();
-    let stake_state: StakeStateV2 = stake_account.state().unwrap();
-    let StakeStateV2::Stake(_, stake, _) = stake_state else {
-        panic!()
-    };
-    check_activation_status(&stake.delegation, StakeActivationState::Inactive, 0);
-
-    // `stake2_keypair` should now be delegated to `vote2_keypair` and fully activated
-    let stake2_account = rpc_client.get_account(&stake2_keypair.pubkey()).unwrap();
-    let stake2_state: StakeStateV2 = stake2_account.state().unwrap();
-
-    match stake2_state {
-        StakeStateV2::Stake(meta, stake, _) => {
-            assert_eq!(stake.delegation.voter_pubkey, vote2_keypair.pubkey());
-            check_activation_status(
-                &stake.delegation,
-                StakeActivationState::Active,
-                50_000_000_000 - meta.rent_exempt_reserve,
-            );
-        }
-        _ => panic!("Unexpected stake2 state!"),
-    };
-}
 
 #[test]
 fn test_stake_delegation_force() {
@@ -396,7 +131,6 @@ fn test_stake_delegation_force() {
         nonce_authority: 0,
         memo: None,
         fee_payer: 0,
-        redelegation_stake_account: None,
         compute_unit_price: None,
     };
     process_command(&config).unwrap();
@@ -440,7 +174,6 @@ fn test_stake_delegation_force() {
         nonce_authority: 0,
         memo: None,
         fee_payer: 0,
-        redelegation_stake_account: None,
         compute_unit_price: None,
     };
     process_command(&config).unwrap_err();
@@ -458,7 +191,6 @@ fn test_stake_delegation_force() {
         nonce_authority: 0,
         memo: None,
         fee_payer: 0,
-        redelegation_stake_account: None,
         compute_unit_price: None,
     };
     process_command(&config).unwrap();
@@ -537,7 +269,6 @@ fn test_seed_stake_delegation_and_deactivation() {
         nonce_authority: 0,
         memo: None,
         fee_payer: 0,
-        redelegation_stake_account: None,
         compute_unit_price: None,
     };
     process_command(&config_validator).unwrap();
@@ -629,7 +360,6 @@ fn test_stake_delegation_and_deactivation() {
         nonce_authority: 0,
         memo: None,
         fee_payer: 0,
-        redelegation_stake_account: None,
         compute_unit_price: None,
     };
     process_command(&config_validator).unwrap();
@@ -745,7 +475,6 @@ fn test_offline_stake_delegation_and_deactivation() {
         nonce_authority: 0,
         memo: None,
         fee_payer: 0,
-        redelegation_stake_account: None,
         compute_unit_price: None,
     };
     config_offline.output_format = OutputFormat::JsonCompact;
@@ -768,7 +497,6 @@ fn test_offline_stake_delegation_and_deactivation() {
         nonce_authority: 0,
         memo: None,
         fee_payer: 0,
-        redelegation_stake_account: None,
         compute_unit_price: None,
     };
     process_command(&config_payer).unwrap();
@@ -906,7 +634,6 @@ fn test_nonced_stake_delegation_and_deactivation() {
         nonce_authority: 0,
         memo: None,
         fee_payer: 0,
-        redelegation_stake_account: None,
         compute_unit_price: None,
     };
     process_command(&config).unwrap();