+++ sdk/program/src/stake/instruction.rs
@@ -1,5 +1,8 @@
-#[allow(deprecated)]
-use crate::stake::config;
+// Remove the following `allow` when the `Redelegate` variant is renamed to
+// `Unused` starting from v3.
+// Required to avoid warnings from uses of deprecated types during trait derivations.
+#![allow(deprecated)]
+
 use {
     crate::{
         clock::{Epoch, UnixTimestamp},
@@ -7,6 +10,7 @@ use {
         program_error::ProgramError,
         pubkey::Pubkey,
         stake::{
+            config,
             program::id,
             state::{Authorized, Lockup, StakeAuthorize, StakeStateV2},
         },
@@ -306,6 +310,7 @@ pub enum StakeInstruction {
     ///   3. `[]` Unused account, formerly the stake config
     ///   4. `[SIGNER]` Stake authority
     ///
+    #[deprecated(since = "2.1.0", note = "Redelegate will not be enabled")]
     Redelegate,
 
     /// Move stake between accounts with the same authorities and lockups, using Staker authority.
@@ -726,7 +731,6 @@ pub fn delegate_stake(
         AccountMeta::new_readonly(*vote_pubkey, false),
         AccountMeta::new_readonly(sysvar::clock::id(), false),
         AccountMeta::new_readonly(sysvar::stake_history::id(), false),
-        #[allow(deprecated)]
         // For backwards compatibility we pass the stake config, although this account is unused
         AccountMeta::new_readonly(config::id(), false),
         AccountMeta::new_readonly(*authorized_pubkey, true),
@@ -832,7 +836,6 @@ fn _redelegate(
         AccountMeta::new(*stake_pubkey, false),
         AccountMeta::new(*uninitialized_stake_pubkey, false),
         AccountMeta::new_readonly(*vote_pubkey, false),
-        #[allow(deprecated)]
         // For backwards compatibility we pass the stake config, although this account is unused
         AccountMeta::new_readonly(config::id(), false),
         AccountMeta::new_readonly(*authorized_pubkey, true),
@@ -840,6 +843,7 @@ fn _redelegate(
     Instruction::new_with_bincode(id(), &StakeInstruction::Redelegate, account_metas)
 }
 
+#[deprecated(since = "2.1.0", note = "Redelegate will not be enabled")]
 pub fn redelegate(
     stake_pubkey: &Pubkey,
     authorized_pubkey: &Pubkey,
@@ -858,6 +862,7 @@ pub fn redelegate(
     ]
 }
 
+#[deprecated(since = "2.1.0", note = "Redelegate will not be enabled")]
 pub fn redelegate_with_seed(
     stake_pubkey: &Pubkey,
     authorized_pubkey: &Pubkey,