+++ transaction-status/src/parse_stake.rs
@@ -284,6 +284,7 @@ pub fn parse_stake(
                 }),
             })
         }
+        #[allow(deprecated)]
         StakeInstruction::Redelegate => {
             check_num_stake_accounts(&instruction.accounts, 5)?;
             Ok(ParsedInstructionEnum {